Job now tracks time {enqueued, started, finished}
The idea is that the web frontend can use it to display how long it's been
waiting in the queue, for how long its been running and finally how long the job
took to finish.

I've opted to implement the new Job as vals and not vars. When one of the timing
fields change we do so by copying the object. One reason is that the Job object
is already copied around actors and that makes it tricker to keep track of their
internal states as well - easier to just keep replacing the entire object.

The reason for the specialized json serializer is that we don't want the time
fields when creating the object from JSON - which is done by the '/enqueue/'
URL.