Modify the way npm is integrated in sbt
This brings it in line with the similar changes in the data project. It makes
"npm publish" a direct dependency of "assembly"; this way it's no longer
possible to build a broken assembly jar file (i.e one with frontend files
missing).

It relies on the new "publish" gulp target.