Start conversion of entire stilt backend
Instead of running a few large simulations we want to run a lot of very small
ones. We do this by first calculating the number of "slots" in a stilt job,
where each slot is a 3-hour window in the entire simulation. Each slot then
becomes a small simulation. These small simulations can the be run independently
and then merged back together
together.