Fix broken cache-hit detection in SlotArchive
It used to be that the cache worked by just looking for directories, i.e
'/stiltweb/slots/station/month/slot'.

But when we added the saving of csv files to the slot directories, we had a lot
of slot directories, and might still have, that had three of the files but not
the csv files. So then the cache-hit detector started considering only slot
directories with the four needed files as "available".

The bug was that the algorithm look for files using the old-style stilt names,
that is, it looked for a file named '.RDatablablablabl' when it should have been
looking for 'rdata'.